+++ src/__stories__/image-story.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import {Image, Stack, Text2, Text3, useTheme} from '..';
+import {Image, Stack, Text2, Text3, Title1, useTheme} from '..';
+import {isRunningAcceptanceTest} from '../utils/platform';
 
 export default {
     title: 'Components/Others/Image',
@@ -9,24 +10,42 @@ export const Default: StoryComponent = () => {
     const {colors} = useTheme();
     return (
         <Stack space={16} dataAttributes={{testid: 'image-story'}}>
+            <Title1>aspectRatio: 1:1</Title1>
             <Image src="https://i.imgur.com/G8aJDhB.jpeg" noBorderRadius aspectRatio="1:1" width={200} />
+            <Title1>aspectRatio: 4:3</Title1>
             <Image src="https://i.imgur.com/G8aJDhB.jpeg" aspectRatio="4:3" width={200} />
+            <Title1>aspectRatio: 7:10</Title1>
             <Image src="https://i.imgur.com/G8aJDhB.jpeg" aspectRatio="7:10" width={200} />
+            <Title1>aspectRatio: 16:9</Title1>
             <Image src="https://i.imgur.com/G8aJDhB.jpeg" aspectRatio="16:9" width={200} />
+            <Title1>width: 230; height: 100</Title1>
             <Image src="https://i.imgur.com/G8aJDhB.jpeg" width={230} height={100} />
             <div
                 style={{
                     resize: 'both',
                     overflow: 'auto',
-                    width: 300,
+                    width: 320,
                     border: `1px solid ${colors.border}`,
                     padding: 16,
                 }}
             >
                 <Stack space={16}>
-                    <Image src="https://i.imgur.com/G8aJDhB.jpeg" aspectRatio={16 / 9} />
-                    <Image src="https://i.imgur.com/G8aJDhB.jpeg" width="50%" aspectRatio={16 / 9} />
+                    <Title1>width: 50%; aspectRatio: 4:3</Title1>
                     <Image src="https://i.imgur.com/G8aJDhB.jpeg" width="50%" aspectRatio={4 / 3} />
+
+                    {!isRunningAcceptanceTest() && (
+                        // for some reason, adding this image to the screenshot test makes it unstable
+                        <Stack space={16}>
+                            <Title1>width: 50%; aspectRatio: 0</Title1>
+                            <Image src="https://i.imgur.com/G8aJDhB.jpeg" width="50%" aspectRatio={0} />
+                        </Stack>
+                    )}
+
+                    <Title1>width: 100%; aspectRatio: 4:3</Title1>
+                    <Image src="https://i.imgur.com/G8aJDhB.jpeg" aspectRatio={4 / 3} />
+
+                    <Title1>width: 100%; aspectRatio: 0</Title1>
+                    <Image src="https://i.imgur.com/G8aJDhB.jpeg" aspectRatio={0} />
                 </Stack>
             </div>
             <Stack space={16}>