+++ src/__stories__/video-story.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import {Stack, ButtonPrimary, Inline, Title1, Video} from '..';
+import {Stack, ButtonPrimary, Inline, Title2, Video, Title1, Text3} from '..';
 
 export default {
     title: 'Components/Others/Video',
@@ -14,14 +14,31 @@ export const Default: StoryComponent = () => {
     const videoRefWithPoster = React.useRef<HTMLVideoElement | null>(null);
 
     return (
-        <Stack space={64} data-testid="video">
-            <Stack space={16}>
-                <Title1>Auto Play</Title1>
-                <Video width={480} height={480 / 2.4} src={VIDEO_SRC} />
+        <Stack space={64} dataAttributes={{testid: 'video'}}>
+            <Stack space={8}>
+                <Title2>Video component issues/limitations</Title2>
+                <Text3 regular as="p">
+                    1. Using automatic aspect ratio (`aspectRatio={0}`), if the poster and video have
+                    different aspect ratios, the video size will resize when playing/stopping the video. This
+                    is how the HTML `video` element works and happens in all browsers.
+                </Text3>
+                <Text3 regular as="p">
+                    2. In Safari, with automatic aspect ratio. On the first video reproduction a strange
+                    resize happens. After stoping and playing the same video, it works as expected.
+                </Text3>
+                <Text3 regular as="p">
+                    To avoid these issues always try to use a specific aspect ratio or a fixed width and
+                    height.
+                </Text3>
             </Stack>
 
             <Stack space={16}>
-                <Title1>Manual Play</Title1>
+                <Title2>Auto Play (default)</Title2>
+                <Video width={480} height={480 / 2.4} src={VIDEO_SRC} poster={POSTER_SRC} />
+            </Stack>
+
+            <Stack space={16}>
+                <Title2>Manual Play (without Poster)</Title2>
                 <Video
                     width={480}
                     height={480 / 2.4}
@@ -50,7 +67,7 @@ export const Default: StoryComponent = () => {
             </Stack>
 
             <Stack space={16}>
-                <Title1>Manual Play with Poster</Title1>
+                <Title2>Manual Play (with Poster)</Title2>
                 <Video
                     poster={POSTER_SRC}
                     width={480}
@@ -84,15 +101,23 @@ export const Default: StoryComponent = () => {
             </Stack>
 
             <Stack space={16}>
-                <Title1>Aspect ratio</Title1>
-                <Video src={VIDEO_SRC} width={480} aspectRatio="16:9" />
-                <Video src={VIDEO_SRC} width={480} aspectRatio="4:3" />
+                <Title2>Aspect ratio</Title2>
+                <Title1>16:9</Title1>
+                <Video src={VIDEO_SRC} poster={POSTER_SRC} width={480} aspectRatio="16:9" />
+                <Title1>4:3</Title1>
+                <Video src={VIDEO_SRC} poster={POSTER_SRC} width={480} aspectRatio="4:3" />
+                <Title1>0 (keep original aspect ratio)</Title1>
+                <Video src={VIDEO_SRC} poster={POSTER_SRC} width={480} aspectRatio={0} />
             </Stack>
 
             <Stack space={16}>
-                <Title1>Aspect ratio width 100%</Title1>
-                <Video src={VIDEO_SRC} aspectRatio="16:9" />
-                <Video src={VIDEO_SRC} aspectRatio="4:3" />
+                <Title2>Aspect ratio width 100%</Title2>
+                <Title1>16:9</Title1>
+                <Video width="100%" src={VIDEO_SRC} poster={POSTER_SRC} aspectRatio="16:9" />
+                <Title1>4:3</Title1>
+                <Video width="100%" src={VIDEO_SRC} poster={POSTER_SRC} aspectRatio="4:3" />
+                <Title1>0 (keep original aspect ratio)</Title1>
+                <Video width="100%" src={VIDEO_SRC} poster={POSTER_SRC} aspectRatio={0} />
             </Stack>
         </Stack>
     );