+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## [11.12.1](https://github.com/Telefonica/mistica-web/compare/v11.12.0...v11.12.1) (2022-06-24)
+
+
+### Bug Fixes
+
+* **Video:** Element size problem when playing in Safari ([bacdae1](https://github.com/Telefonica/mistica-web/commit/bacdae172a9e0399a4d738de0e92c74bdb0fa295))
+
+# [11.12.0](https://github.com/Telefonica/mistica-web/compare/v11.11.1...v11.12.0) (2022-06-23)
+
+
+### Bug Fixes
+
+* **Video:** problem with aspectRatio 0 in Safari ([#498](https://github.com/Telefonica/mistica-web/issues/498)) ([24452eb](https://github.com/Telefonica/mistica-web/commit/24452eb50a4d7335853b4c0f76747f7c229ce62c))
+
+
+### Features
+
+* **Skin:** allow to customize font weight of some text presets ([#495](https://github.com/Telefonica/mistica-web/issues/495)) ([cf5968c](https://github.com/Telefonica/mistica-web/commit/cf5968cca99eca56533fac4df5dec6f3ca95042a))
+
 ## [11.11.1](https://github.com/Telefonica/mistica-web/compare/v11.11.0...v11.11.1) (2022-06-22)